+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/database/DatabaseIpConfig.kt
@@ -1,0 +1,25 @@
+package com.atlassian.performance.tools.infrastructure.api.database
+
+import com.atlassian.performance.tools.infrastructure.api.Sed
+import com.atlassian.performance.tools.infrastructure.api.jira.install.hook.PostInstallHooks
+import com.atlassian.performance.tools.infrastructure.api.jira.install.InstalledJira
+import com.atlassian.performance.tools.infrastructure.api.jira.install.hook.PostInstallHook
+import com.atlassian.performance.tools.ssh.api.SshConnection
+
+class DatabaseIpConfig(
+    private val databaseIp: String
+) : PostInstallHook {
+
+    override fun call(
+        ssh: SshConnection,
+        jira: InstalledJira,
+        hooks: PostInstallHooks
+    ) {
+        Sed().replace(
+            connection = ssh,
+            expression = "(<url>.*(@(//)?|//))" + "([^:/]+)" + "(.*</url>)",
+            output = """\1$databaseIp\5""",
+            file = "${jira.home}/dbconfig.xml"
+        )
+    }
+}