+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/install/JiraInstallation.kt
@@ -1,0 +1,13 @@
+package com.atlassian.performance.tools.infrastructure.api.jira.install
+
+import com.atlassian.performance.tools.ssh.api.SshConnection
+import net.jcip.annotations.ThreadSafe
+
+@ThreadSafe
+interface JiraInstallation {
+
+    fun install(
+        ssh: SshConnection,
+        server: TcpServer
+    ): InstalledJira
+}