+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/install/TcpServer.kt
@@ -1,0 +1,19 @@
+package com.atlassian.performance.tools.infrastructure.api.jira.install
+
+class TcpServer(
+    val ip: String,
+    val publicPort: Int,
+    val privatePort: Int,
+    val name: String
+) {
+    constructor(
+        ip: String,
+        port: Int,
+        name: String
+    ) : this(
+        ip,
+        port,
+        port,
+        name
+    )
+}