+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/jira/install/hook/HookedJiraInstallation.kt
@@ -1,0 +1,22 @@
+package com.atlassian.performance.tools.infrastructure.api.jira.install.hook
+
+import com.atlassian.performance.tools.infrastructure.api.jira.install.InstalledJira
+import com.atlassian.performance.tools.infrastructure.api.jira.install.JiraInstallation
+import com.atlassian.performance.tools.infrastructure.api.jira.install.TcpServer
+import com.atlassian.performance.tools.ssh.api.SshConnection
+
+class HookedJiraInstallation(
+    private val installation: JiraInstallation,
+    private val hooks: PreInstallHooks
+) : JiraInstallation {
+
+    override fun install(
+        ssh: SshConnection,
+        server: TcpServer
+    ): InstalledJira {
+        hooks.call(ssh, server)
+        val installed = installation.install(ssh, server)
+        hooks.postInstall.call(ssh, installed)
+        return installed
+    }
+}