+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/api/virtualusers/SshVirtualUsers.kt
@@ -89,8 +89,4 @@ class SshVirtualUsers(
             }
         }
     }
-
-    override fun toString(): String {
-        return "SshVirtualUsers(name='$name', nodeOrder=$nodeOrder, resultsTransport=$resultsTransport, jarName='$jarName', ssh=$ssh, logger=$logger, jdk=$jdk)"
-    }
 }