+++ src/main/kotlin/com/atlassian/performance/tools/infrastructure/ChromedriverInstaller.kt
@@ -2,6 +2,8 @@ package com.atlassian.performance.tools.infrastructure
 
 import com.atlassian.performance.tools.infrastructure.api.os.Ubuntu
 import com.atlassian.performance.tools.ssh.api.SshConnection
+import org.apache.http.client.methods.HttpGet
+import org.apache.http.impl.client.HttpClients
 import java.net.URI
 import java.time.Duration
 
@@ -19,5 +21,15 @@ internal class ChromedriverInstaller(private val uri: URI) {
                 Ubuntu().install(ssh, listOf("zip", "libglib2.0-0", "libnss3"), Duration.ofMinutes(2))
             }
         )
+    }
+
+    companion object {
+        fun getLatestVersion(minorVersion: String?): String {
+            val httpclient = HttpClients.createDefault()
+            val versionSuffix = if (minorVersion != null) "_$minorVersion" else "";
+            val get = HttpGet("https://chromedriver.storage.googleapis.com/LATEST_RELEASE$versionSuffix")
+            val response = httpclient.execute(get)
+            return response.entity.content.bufferedReader().use { it.readLine() }
+        }
     }
 }