+++ CHANGELOG.md
@@ -24,7 +24,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/infrastructure/compare/release-4.17.5...master
+[Unreleased]: https://github.com/atlassian/infrastructure/compare/release-4.18.0...master
+
+## [4.18.0] - 2021-04-14
+[4.18.0]: https://github.com/atlassian/infrastructure/compare/release-4.17.5...release-4.18.0
 
 ### Added
 - Point to remote files on SSH hosts via `RemotePath`.